use CMake_OpenCV.txt for all projects, and see more below
do not set OpenCV_DIR, let user provide it
tested with static build of OpenCV with MSVC
Jpeg add CMakeLists.txt
fix warnings
cmake do not manually add other libraries, instead add_subdirectory
add suffix to library names
fix legacy opencv code

+++ Calibrate/main_calibrate.cpp
@@ -36,7 +36,7 @@ int main(int argc, char *argv[])
     // save object & image points to external files
     CHECK(cal.save_chessboard(cap, delay));
 
-    cv::Size sz(cap.get(cv::CAP_PROP_FRAME_WIDTH), cap.get(cv::CAP_PROP_FRAME_HEIGHT));
+    cv::Size sz(static_cast<int>(cap.get(cv::CAP_PROP_FRAME_WIDTH)), static_cast<int>(cap.get(cv::CAP_PROP_FRAME_HEIGHT)));
 
     // read object & image points from external files
     std::vector<std::vector<cv::Point3f>> object_points;
